+++ .gitignore
@@ -247,6 +247,10 @@ ClientBin/
 *.publishsettings
 orleans.codegen.cs
 
+#KDevelop
+*.kdev4
+.kdev4/*
+
 # Including strong name files can present a security risk
 # (https://github.com/github/gitignore/pull/2483#issue-259490424)
 #*.snk
