+++ CMakeLists.txt
@@ -5,7 +5,7 @@ cmake_minimum_required(VERSION 3.15)
 #
 
 project(
-  "Project"
+  "NewName"
   VERSION 0.1.0 
   LANGUAGES CXX
 )
