+++ cmake/NewNameConfig.cmake.in
