+++ src/tmp.cpp
@@ -1,3 +1,3 @@
-#include "project/tmp.hpp"
+#include "NewName/tmp.hpp"
 
 int tmp::add(int a, int b) { return a + b; }
