+++ test/src/tmp_test.cpp
@@ -1,4 +1,4 @@
-#include "project/tmp.hpp"
+#include "NewName/tmp.hpp"
 
 #include <gtest/gtest.h>
